get chain for ff layer fixed float to int.

+++ Network/layers/BatchNormalizationLayer.cpp
@@ -37,6 +37,6 @@ void cn::BatchNormalizationLayer::run(const cn::Bitmap<float> &bitmap) {
     Layer::run(bitmap);
 }
 
-float cn::BatchNormalizationLayer::getChain(const Vector3<float> &input) {
+float cn::BatchNormalizationLayer::getChain(const Vector3<int> &input) {
     return 0;
 }
